feat: crawler benchmark

+++ crawler/getUrls.js
@@ -1,5 +1,5 @@
 const axios = require('axios');
-const cheerio = require('cheerio');
+const jsdom = require('jsdom');
 const https = require('https');
 const fs = require('fs');
 const path = require('path');
@@ -10,26 +10,31 @@ const path = require('path');
  * For example, to set MAX_URLS to 2000, run: `node getUrls.js 2000`
  * If no command line argument is provided, the default value is 100k.
  */
-const MAX_URLS = process.argv[2] || 100 * 1000;
+const MAX_URLS = process.argv[2] || 200 * 1000;
 const startingUrl = 'https://atlas.cs.brown.edu/data/gutenberg/';
 const visitedUrls = new Set();
-const urlPath = path.join(__dirname, 'datasets.txt');
+const datasetsUrl = `datasets-${MAX_URLS/2}.txt`;
+const urlPath = path.join(__dirname, datasetsUrl);
 
 const agent = new https.Agent({
   rejectUnauthorized: false,
 });
+
 async function crawl(url) {
   if (visitedUrls.size >= MAX_URLS) {
     return;
   }
 
   try {
-    const response = await axios.get(url, {httpsAgent: agent});
-    const $ = cheerio.load(response.data);
+    const response = await axios.get(url, { httpsAgent: agent });
+    const dom = new jsdom.JSDOM(response.data);
     const curUrls = [];
-    visitedUrls.add(url);
-    $('a').each((index, element) => {
-      const href = $(element).attr('href');
+    curUrls.forEach((curUrl) => {
+      visitedUrls.add(curUrl);
+      fs.appendFileSync(urlPath, curUrl + '\n');
+    });
+    dom.window.document.querySelectorAll("a").forEach((element) => {
+      const href = element.getAttribute('href');
       if (href) {
         const resolvedUrl = new URL(href, url).href.split('?')[0];
         const cleanUrl = resolvedUrl.split('#')[0];
@@ -51,20 +56,45 @@ async function crawl(url) {
   }
 }
 
-crawl(startingUrl).then(() => {
-  console.log(`Found ${visitedUrls.size} unique URLs:`);
-
-  fs.writeFile(
-    urlPath,
-    Array.from(visitedUrls)
-      .filter((url) => url.endsWith('.txt'))
-      .join('\n'),
-    (err) => {
-      if (err) {
-        console.error('Error writing to file:', err);
-      } else {
-        console.log('Visited URLs written to datasets.txt');
-      }
-    },
-  );
-});
+if (fs.existsSync(urlPath)) {
+  const data = fs.readFileSync(urlPath, 'utf8');
+  const urls = data.split('\n');
+  urls.forEach((url) => {
+    visitedUrls.add(url);
+  });
+  crawl(urls.at(-1).then(()=>{
+    console.log(`Found ${visitedUrls.size} unique URLs:`);
+  
+    fs.writeFile(
+      urlPath,
+      Array.from(visitedUrls)
+        .filter((url) => url.endsWith('.txt'))
+        .join('\n'),
+      (err) => {
+        if (err) {
+          console.error('Error writing to file:', err);
+        } else {
+          console.log('Visited URLs written to ' + urlPath);
+        }
+      },
+    );
+  }))
+} else {
+  crawl(startingUrl).then(() => {
+    console.log(`Found ${visitedUrls.size} unique URLs:`);
+  
+    fs.writeFile(
+      urlPath,
+      Array.from(visitedUrls)
+        .filter((url) => url.endsWith('.txt'))
+        .join('\n'),
+      (err) => {
+        if (err) {
+          console.error('Error writing to file:', err);
+        } else {
+          console.log('Visited URLs written to ' + urlPath);
+        }
+      },
+    );
+  });
+}
